Rewritten constructor + copy handling
Before this change we had a bunch of independent constructor functions which
looked like "%IO* @IO$construct()" and the plan was to implement Object.copy
by having each class override that method.

This was completely broken because:
 - Constructors can't be called in chains (child class calls parent) because
   the size of the class object was hardcoded. There was already a failed
   attempt to do chain calls in the runtime file!
 - new SELF_TYPE is not implementable since the constructor function is
   needed at runtime.
 - Object.copy would give the wrong result if a user class overrides copy
   and calls X@Object.copy.

So these have been added to Object's vtable:
	void (%Object*)*,          ; Constructor
	void (%Object*, %Object*)*,; Copy constructor

The constructor initializes the attributes in a class. Code now calls
@new_object to create + initialize objects.

The copy constructor does the same but copies values from the second
argument into the first. The Object.copy implementation calls this
method (no generated code does).