cool_program: fix use-after-free concerning method slot destruction
The case arises if a cool_method owning a cool_method_slot is destructed
before all of the overriding methods. This can easily happen since the
order of destruction is not specified. In this case, the overriding
methods try to dereference _slot in their destructors after it is freed.

Fix by adding a _slot_owner field to cool_method so the method does not
need to dereference _slot to determin if it's the owner.